+++ docs/dev/CHANGELOG.md
@@ -52,6 +52,7 @@ Released on: ???
 - Fixed some baits being added to Fishing profit tracker after adding Bait Bag preview in slot #9.
 - Fixed Magma Lord Necklace crafting recipe for /feeshGearCraftPrices.
 - Fixed some overlays data not auto-resetting after disabling that overlay.
+- Fixed trophy armor checks to be executed more often than needed.
 
 # 1.6.1